fix: 미들웨어에서 axios 대신 fetch를 사용하도록 변경
